Update conclusion.tex
Hey Tim, good work. One observation on the use of terms. In a few places in the text you say 'infrastructureless software' which is an oxymoron. Software is an infrastructure. I know you probably mean infrastructureless communication, but in my opinion it is also a bit gimmicky marketing thing which will not hold the test of time, just like 'ambient intelligence'. Maybe you would be better off saying 'server-less' and 'ad-hoc' communication.